Tag the frontend interface by mac address rather than by name
After deploying a cluster it was found that floating IPs were not
working despite being deployed and configured within qfsd. It turned out
the frontend interface had a different name. With this, we use the mac
address of the first interface listed in the instance metadata service
rather than assuming the name will always be the same.